Griffin/backend/requirements.txt
python-dotenv
pydantic
fastapi
uvicorn
python-multipart